+++ CHANGES.rst
@@ -18,6 +18,8 @@ Unreleased
 -   Stop support for python 3.7 :pr:`794`
 -   Added shorter format to :class:`~fields.WeekField`
     defaults :pr:`765`
+-   Implement :class:`~validators.ReadOnly` and
+    :class:`~validators.Disabled `:pr:`788`
 
 Version 3.0.1
 -------------